refactor(user): move request and response DTOs to respective packages
